requirements.txt
Flask==2.3.3
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-mock==3.14.0